+++ anki_addon/ui/word_info.py
@@ -1,3 +1,4 @@
+import anki.errors
 from aqt.qt import (
     Qt, QSizePolicy,
     QWidget, QHBoxLayout, QVBoxLayout, QLabel, QPushButton,
@@ -20,6 +21,7 @@ class WordInfoPanel(QGroupBox):
         self.search_info = search_info
         self.word_info = None
         self.translations = None
+        self.note = None
         self._note_type_id = None
         self._sonaveeb = sonaveeb
         # Add status label
@@ -45,13 +47,24 @@ class WordInfoPanel(QGroupBox):
         data_layout.addWidget(self._pos_label)
         data_layout.addWidget(self._translations_label)
         data_layout.addWidget(self._translations_status)
-        self._add_button = QPushButton()
-        self._add_button.setEnabled(False)
-        self._add_button.setSizePolicy(QSizePolicy.Policy.Maximum, QSizePolicy.Policy.Maximum)
+
+        self._add_button = QPushButton('Add')
+        self._add_button.setFixedWidth(100)
+        self._add_button.hide()
         self._add_button.clicked.connect(self.add_button_clicked)
+        self._delete_button = QPushButton('Delete')
+        self._delete_button.setFixedWidth(100)
+        self._delete_button.hide()
+        self._delete_button.clicked.connect(self.delete_button_clicked)
+        self._replace_button = QPushButton('Replace')
+        self._replace_button.setFixedWidth(100)
+        self._replace_button.hide()
+        self._replace_button.clicked.connect(self.replace_button_clicked)
         buttons_layout = QVBoxLayout()
         buttons_layout.setAlignment(Qt.AlignmentFlag.AlignTop)
         buttons_layout.addWidget(self._add_button)
+        buttons_layout.addWidget(self._delete_button)
+        buttons_layout.addWidget(self._replace_button)
         content_layout = QHBoxLayout()
         content_layout.addLayout(data_layout)
         content_layout.addLayout(buttons_layout)
@@ -70,33 +83,41 @@ class WordInfoPanel(QGroupBox):
         self.setSizePolicy(QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Maximum)
 
         # Request word info
-        self.set_note_added(False)
         self.request_word_info()
 
     def set_translations(self, translations, external=False, limit=3):
-        filtered = [t.strip('!., ') for t in translations[:limit]]
-        self.translations = ', '.join(filtered)
-        self._translations_label.setText(self.translations)
-        self._translations_label.show()
-        self._translations_status.setText('Google Translated')
-        self._translations_status.setVisible(external)
+        if len(translations) == 0:
+            self._translations_status.setText('Translations unavailable')
+            self._translations_status.show()
+            self._add_button.setEnabled(False)
+            self._replace_button.hide()
+        else:
+            filtered = [t.strip('!., ') for t in translations[:limit]]
+            self.translations = ', '.join(filtered)
+            self._translations_label.setText(self.translations)
+            self._translations_label.show()
+            self._translations_status.setText('Google Translated')
+            self._translations_status.setVisible(external)
+            self._add_button.setEnabled(True)
+            self.check_note_identical()
 
     def set_translation_language(self, lang):
         self.lang = lang
         self._translations_label.hide()
-        if self.word_info is None:
-            return
-        translations = self.word_info.lexemes[0].translations.get(self.lang)
-        if translations is not None:
-            self.set_translations(translations, False)
-        elif len(self.word_info.lexemes[0].translations) == 0:
-            self.translations_received([])
-        else:
-            self.request_cross_translations()
+        self._add_button.setEnabled(False)
+        self._replace_button.hide()
+        if self.word_info is not None:
+            translations = self.word_info.lexemes[0].translations.get(self.lang)
+            if translations is not None:
+                self.set_translations(translations, False)
+            elif len(self.word_info.lexemes[0].translations) == 0:
+                self.set_translations([])
+            else:
+                self.request_cross_translations()
 
     def set_deck(self, deck_id):
         self.deck_id = deck_id
-        self.check_note_added()
+        self.check_note_exists()
 
     def set_status(self, status):
         self._status_label.setText(status)
@@ -111,14 +132,26 @@ class WordInfoPanel(QGroupBox):
         self.set_translation_language(self.lang)
         self._stack.setCurrentWidget(self._content)
 
-    def set_note_added(self, state):
-        self._add_button.setText('Note added' if state else 'Add note')
-        self._add_button.setEnabled(not state and self.translations is not None)
-
-    def check_note_added(self):
+    def check_note_exists(self):
         deck = mw.col.decks.get(self.deck_id)['name']
         existing_notes = mw.col.find_notes(f'URL:"{self.word_info.url}" deck:"{deck}"')
-        self.set_note_added(len(existing_notes) != 0)
+        exists = len(existing_notes) > 0
+        self._add_button.setVisible(not exists)
+        self._delete_button.setVisible(exists)
+        if exists:
+            self.note = mw.col.get_note(existing_notes[0])
+            self.check_note_identical()
+        else:
+            self.note = None
+
+    def check_note_identical(self):
+        if self.note is not None:
+            identical = (
+                self.note['Morphology'] == self.word_info.short_record() and
+                self.note['Translation'] == self.translations and
+                self.note['URL'] == self.word_info.url
+            )
+            self._replace_button.setVisible(not identical)
 
     def get_note_type_id(self):
         if self._note_type_id is not None:
@@ -129,7 +162,9 @@ class WordInfoPanel(QGroupBox):
                 self._note_type_id = ntid
                 return ntid
             else:
-                QMessageBox.warning(self, 'Malformed note type')
+                QMessageBox.warning(self, 'Malformed note type',
+                    f'Note type "{mw.col.models.get(ntid)["name"]}" is malformed.'
+                    ' You can try to remove it from "Tools -> Manage Note Types"')
 
         self._note_type_id = add_note_type()
         return self._note_type_id
@@ -137,13 +172,31 @@ class WordInfoPanel(QGroupBox):
     def add_note(self):
         if (ntid := self.get_note_type_id()) is not None:
             note = mw.col.new_note(ntid)
-            note['Morphology'] = self.word_info.short_record()
-            note['Translation'] = self.translations
-            note['URL'] = self.word_info.url
-            if self.word_info.pos is not None:
-                note.add_tag(self.word_info.pos)
+            self.fill_note(note)
             mw.col.add_note(note, self.deck_id)
-            self.set_note_added(True)
+            self.note = note
+
+    def update_note(self):
+        if self.note is not None:
+            self.fill_note(self.note)
+            mw.col.update_note(self.note, self.deck_id)
+
+    def delete_note(self):
+        if self.note is not None:
+            results = mw.col.remove_notes([self.note.id])
+            if results.count == 0:
+                raise RuntimeError(
+                    'Your database appears to be in an inconsistent state.'
+                    ' Please use the Check Database action.')
+            else:
+                self.note = None
+
+    def fill_note(self, note):
+        note['Morphology'] = self.word_info.short_record()
+        note['Translation'] = self.translations
+        note['URL'] = self.word_info.url
+        if self.word_info.pos is not None:
+            note.add_tag(self.word_info.pos)
 
     def request_word_info(self):
         self.set_status('Loading...')
@@ -155,7 +208,6 @@ class WordInfoPanel(QGroupBox):
         operation.run_in_background()
 
     def request_cross_translations(self):
-        self._add_button.setEnabled(False)
         self._translations_status.setText('Google translating...')
         self._translations_status.show()
         operation = QueryOp(
@@ -179,23 +231,37 @@ class WordInfoPanel(QGroupBox):
 
     def word_info_received(self, word_info):
         if word_info is None:
-            self._add_button.setEnabled(False)
             self.set_status('Failed to obtain word info :(')
         else:
             self.set_word_info(word_info)
-            self.check_note_added()
+            self.check_note_exists()
 
     def translations_received(self, translations):
-        if len(translations) == 0:
-            self._add_button.setEnabled(False)
-            self._translations_status.setText('Translations unavailable')
-            self._translations_status.show()
-        else:
-            self._add_button.setEnabled(True)
-            # As a special case, add "to" before verbs infinitives in English
-            if self.lang == 'en' and self.word_info.pos == 'tegusõna':
-                translations = [f'to {verb}'.replace('to to ', 'to ') for verb in translations]
-            self.set_translations(translations, True)
+        # As a special case, add "to" before verbs infinitives in English
+        if self.lang == 'en' and self.word_info.pos == 'tegusõna':
+            translations = [f'to {verb}'.replace('to to ', 'to ') for verb in translations]
+        self.set_translations(translations, True)
 
     def add_button_clicked(self):
         self.add_note()
+        self._add_button.hide()
+        self._delete_button.show()
+
+    def delete_button_clicked(self):
+        try:
+            self.delete_note()
+        except RuntimeError as e:
+            QMessageBox.warning(self, 'Failed to delete the note', str(e))
+        else:
+            self._add_button.show()
+            self._delete_button.hide()
+            self._replace_button.hide()
+
+    def replace_button_clicked(self):
+        try:
+            self.update_note()
+        except anki.errors.NotFoundError as e:
+            QMessageBox.warning(self, 'Failed to update the note', str(e))
+        else:
+            self._replace_button.hide()
+