+++ scripts/package.sh
@@ -5,6 +5,6 @@ ADDON_DIR="$REPO_DIR/anki_addon"
 VERSION=$(cd "$REPO_DIR" && git describe)
 
 cd "$ADDON_DIR"
-rm -r __pycache__ meta.json
+rm -rf **/__pycache__ meta.json
 zip -r ../sonaveeb_integration_$VERSION.ankiaddon *