+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Added "Delete" and "Replace" note buttons.
+
+
 ## v0.3.0 - 2024-01-30
 
 - Improved cross-translation algorithm.
@@ -9,11 +14,13 @@
 - Fixed typos in some languages names.
 - Fixed translations refresh when language changes.
 
+
 ## v0.2.0 - 2024-01-28
 
 - Added fallback to Google Translate for missing translations.
 - Added a wide selection of languages for translations.
 - Use system language as default for translations.
+
 
 ## v0.1.0 - 2024-01-27